scrapers/pyproject.toml
[tool.poetry]
name = "scrapers"
version = "0.1.0"
description = ""
authors = ["pajowu <git@ca.pajowu.de>"]

[tool.poetry.dependencies]
python = "^3.8"
beautifulsoup4 = "^4.10.0"
requests = "^2.27.1"
python-docx = "^0.8.11"
xmltodict = "^0.12.0"
more-itertools = "^8.12.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
